refactor: move Reviews related api to own folder
