+++ project/src/store/api-actions.ts
@@ -4,7 +4,7 @@ import { Token } from 'src/services/token';
 import { APIRoute } from 'src/consts/api';
 import { AppRoute } from 'src/consts/consts';
 import { AppDispatch, AppState } from 'src/types/store';
-import { Offer, OfferId, Offers, Reviews } from 'src/types/types';
+import { Offer, OfferId, Offers } from 'src/types/types';
 import { UserAuthorizationData, UserData } from 'src/types/api';
 
 export const checkIfUserAuthorizedAction = createAsyncThunk<UserData, void, {
@@ -52,19 +52,6 @@ export const fetchOffersAction = createAsyncThunk<Offers, undefined, {
   async (_arg, { extra: api }) => {
     const { data: offers } = await api.get<Offers>(APIRoute.Offers);
     return offers;
-  },
-);
-
-export const fetchReviewsAction = createAsyncThunk<Reviews, Offer, {
-  dispatch: AppDispatch;
-  state: AppState;
-  extra: AxiosInstance;
-}>(
-  'data/fetchReviews',
-  async (offer, { extra: api }) => {
-    const url = `${APIRoute.Reviews}${offer.id.toString()}`;
-    const { data: reviews } = await api.get<Reviews>(url);
-    return reviews;
   },
 );
